Splits mech builder UI into multiple files
